Testing and Evaluation (NOT COMPLETED)
I did the "Analyse and Conclude" section and the "Function Requirement Evaulation" part. I started a bit on the non-functional requirement part, and the "Peer Evaulation" part is blank for now. I'll ask hannah and zozo to do it at reccess.

I rewrote some of the functional requirements and purpose part cause i found it was way to hard to code so I had to change some stuff and that's abt it