Finished :D
Added the "Evaluation of your Project in Relation to Testing and Project Management" part.

Finished everything, so there shouldn't be any need to change anything, but I'll see if I find a problem at recess and lunch.

I added red lines in between the sections of "TESING AND EVALUATION" becuase there was just so much, but I couldn't completely seperate them because there were in the same section.

done yayyy